backend/services/service-3/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
kafka-python==2.0.4
six==1.16.0
